Flexible and Coherent Test/Estimation Procedure Based on Restricted Mean Survival Times

Estimates the restricted mean survival time (RMST) with a time window [0, tau], where tau is selected during the procedure from among the time points provided by the user. It also estimates the RMST with a time window [tau1, tau2], where tau1 is selected during the procedure, same as above.

References

Horiguchi M, Cronin A, Takeuchi M, Uno H. A flexible and coherent test/estimation procedure based on restricted mean survival times for censored time-to-event data in randomized clinical trials. Statistics in Medicine 2018. DOI:10.1002/sim.7661.

Horiguchi M, Tian L, Uno H. On assessing survival benefit of immunotherapy using long-term restricted mean survival time. Statistics in Medicine 2023. DOI:10.1002/sim.9662.

See also

Author

Miki Horiguchi, Lu Tian, Hajime Uno

Maintainer: Miki Horiguchi <horiguchimiki@gmail.com>

Examples

#--- Adaptive RMST test/estimation method by rmst2adapt() ---#
data    = rmst2adapt.sample.data()
nmethod = 100 #This is only for example use.
              #Recommended to specify at least 100000 (default) or larger.

a       = rmst2adapt(indata=data, tau_star=seq(6,12,2), method="perturbation",
               nmethod=nmethod, test="2_side", seed=123)
print(a)
#> 
#> <Test result> 
#> Candidate values of tau1: 6 8 10 12 
#> 
#> P-value: 0.85 
#> 
#> 
#> <Treatment effect estimation> 
#> Selected tau: 6 
#> 
#> RMST(arm1) RMST(arm0) 
#>      4.985      4.885 
#> 
#> RMST(arm1-arm0)      lower 0.95      upper 0.95 
#>           0.100          -0.360           0.561 
#> 
#> 

#--- Adaptive long-term RMST test/estimation method by ltrmst2adapt() ---#
data    = cm214_pfs
b       = ltrmst2adapt(indata=data, tau1=c(0,1,2,3), tau2=10, 
               test="2_side", seed=123)
print(b)
#> 
#> RMST over the time window [3,10] 
#> 
#> RMST(arm1) RMST(arm0) 
#>      4.415      4.191 
#> 
#> RMST(arm1-arm0)      lower 0.95      upper 0.95         p_value 
#>           0.224          -0.218           0.666           0.332 
#> 
#>